重构了部分
还有很多未完成

+++ Net/TaskList.cpp
@@ -28,7 +28,7 @@ void TaskList::Clear()
 	taskList_.clear();
 }
 
-int TaskList::Count() const
+size_t TaskList::Count() const
 {
 	return taskList_.size();
 }
